SPAP-25398 Tatt bort bruk av overskytende prosent for de nye tabellene

+++ src/main/java/no/skatteetaten/fastsetting/formueinntekt/forskudd/trekkrutine2024/HeleTabellen.java
@@ -4,9 +4,9 @@ import java.util.LinkedHashMap;
 
 public class HeleTabellen {
     public final LinkedHashMap<Long, Long> alleTrekk;
-    public final int overskytendeProsent;
+    public final double overskytendeProsent;
 
-    public HeleTabellen(LinkedHashMap<Long, Long> alleTrekk, int overskytendeProsent) {
+    public HeleTabellen(LinkedHashMap<Long, Long> alleTrekk, double overskytendeProsent) {
         this.alleTrekk = alleTrekk;
         this.overskytendeProsent = overskytendeProsent;
     }
